RavenDB-26022 fix (document schema): Removed redundant global schema validation toggle logic, cleaned up unused imports, and adjusted UI elements.
RavenDB-26022 feat (document schema): Add global schema validation toggle, license check, and schema playground link
