+++ src/Raven.Studio/typescript/components/pages/database/settings/documentSchema/DocumentSchema.tsx
@@ -57,11 +57,8 @@ import { ValidationSchemaViewSheetPanel } from "components/pages/database/settin
 import { ScriptSyntaxHelp } from "components/pages/database/settings/documentSchema/partials/ScriptSyntaxHelp";
 import { licenseSelectors } from "components/common/shell/licenseSlice";
 import FeatureNotAvailableInYourLicensePopoverBody from "components/common/FeatureNotAvailableInYourLicensePopoverBody";
-import { useAppUrls } from "hooks/useAppUrls";
-import useConfirm from "components/common/ConfirmDialog";
-import { useEventsCollector } from "hooks/useEventsCollector";
-import { StickyHeader } from "components/common/StickyHeader";
 import classNames from "classnames";
+import { StickyHeader } from "components/common/StickyHeader";
 
 const ajv = new Ajv({
     allErrors: true,
@@ -69,7 +66,6 @@ const ajv = new Ajv({
 });
 
 export default function DocumentSchema() {
-    const { forCurrentDatabase: urls } = useAppUrls();
     const dispatch = useAppDispatch();
     const databaseName = useAppSelector(databaseSelectors.activeDatabaseName);
     const { databasesService } = useServices();
@@ -82,17 +78,6 @@ export default function DocumentSchema() {
         return result;
     }, []);
 
-    const allCollectionNames = useAppSelector(documentSchemaSelectors.allCollectionNames);
-    const isGlobalDisabled = useAppSelector(documentSchemaSelectors.isGlobalDisabled);
-    const {
-        value: isTogglingGlobalStatus,
-        setTrue: setIsTogglingGlobalStatusTrue,
-        setFalse: setIsTogglingGlobalStatusFalse,
-    } = useBoolean(false);
-    const confirm = useConfirm();
-    const { reportEvent } = useEventsCollector();
-    const allValidators = useAppSelector(documentSchemaSelectors.allValidators);
-
     const {
         options,
         hasAnyValidator,
@@ -104,54 +89,6 @@ export default function DocumentSchema() {
         setSelectedStatuses,
     } = useDocumentSchema();
 
-    const handleGlobalStatusOperation = async (disabled: boolean) => {
-        const isConfirmed = await confirm({
-            title: disabled
-                ? "Do you want to disable schema validation globally?"
-                : "Do you want to enable schema validation globally?",
-            message: disabled ? (
-                <p>
-                    This will disable schema validation for all collections in the database. Documents will no longer be
-                    validated against their schemas.
-                </p>
-            ) : (
-                <>
-                    <p className="mb-0">
-                        This will <b>enable schema validation globally</b> for all collections in the database.
-                        Documents will be validated against their defined schemas.
-                    </p>
-                    <p>
-                        If a schema has validation <b>disabled individually</b>, you’ll need to{" "}
-                        <b>enable it manually</b>. The global setting does not override per-schema configurations.
-                    </p>
-                </>
-            ),
-            icon: disabled ? "stop" : "play",
-            confirmText: disabled ? "Disable" : "Enable",
-            actionColor: disabled ? "danger" : "success",
-        });
-
-        if (!isConfirmed) {
-            return;
-        }
-
-        try {
-            setIsTogglingGlobalStatusTrue();
-            reportEvent("document-schema", disabled ? "global-disable" : "global-enable");
-
-            dispatch(documentSchemaActions.isGlobalDisabledToggled(disabled));
-
-            await databasesService.saveSchemaValidation(
-                databaseName,
-                documentSchemaUtils.mapToSchemaValidationConfigurationDto(allValidators, disabled)
-            );
-
-            dispatch(documentSchemaActions.validatorsSaved());
-        } finally {
-            setIsTogglingGlobalStatusFalse();
-        }
-    };
-
     return (
         <div className="content-margin">
             <Row className="gy-sm">
@@ -162,68 +99,23 @@ export default function DocumentSchema() {
                         icon="document-schema"
                         licenseBadgeText={hasSchemaValidation ? null : "Professional +"}
                     />
-                    {hasDatabaseAdminAccess && (
-                        <StickyHeader>
-                            <Row>
-                                <div className="d-flex flex-wrap gap-2">
-                                    <div className={hasSchemaValidation ? "" : "item-disabled pe-none"}>
-                                        {hasAnyValidator && (
-                                            <DocumentSchemaFilter
-                                                selectedCollections={selectedCollections}
-                                                setSelectedCollections={handleOnSelectCollection}
-                                                collectionOptions={options}
-                                                selectedStatuses={selectedStatuses}
-                                                setSelectedStatuses={setSelectedStatuses}
-                                                schemasCount={filteredValidators.length}
-                                                isLoading={asyncLoadValidators.loading}
-                                            />
-                                        )}
-                                    </div>
-                                    <div className="d-flex gap-2 align-items-end ms-auto">
-                                        <ConditionalPopover
-                                            conditions={{
-                                                isActive: !hasSchemaValidation,
-                                                message: <FeatureNotAvailableInYourLicensePopoverBody />,
-                                            }}
-                                        >
-                                            <a
-                                                className={classNames("btn btn-secondary rounded-pill", {
-                                                    disabled: !hasSchemaValidation,
-                                                })}
-                                                href={urls.documentSchemaPlayground()}
-                                            >
-                                                <Icon icon="rocket" />
-                                                Schema Playground
-                                            </a>
-                                        </ConditionalPopover>
-                                        {allCollectionNames.length !== 0 && (
-                                            <ConditionalPopover
-                                                conditions={{
-                                                    isActive: !hasSchemaValidation,
-                                                    message: <FeatureNotAvailableInYourLicensePopoverBody />,
-                                                }}
-                                            >
-                                                <ButtonWithSpinner
-                                                    variant={isGlobalDisabled ? "success" : "secondary"}
-                                                    onClick={() => handleGlobalStatusOperation(!isGlobalDisabled)}
-                                                    isSpinning={isTogglingGlobalStatus}
-                                                    disabled={!hasSchemaValidation}
-                                                    className="rounded-pill"
-                                                >
-                                                    <Icon icon={isGlobalDisabled ? "play" : "stop"} />
-                                                    {isGlobalDisabled ? "Enable" : "Disable"} Schema Validation
-                                                </ButtonWithSpinner>
-                                            </ConditionalPopover>
-                                        )}
-                                    </div>
-                                </div>
-                                <div className={classNames("mt-3", !hasSchemaValidation && "item-disabled pe-none")}>
-                                    <DocumentSchemaSelectActions />
-                                </div>
-                            </Row>
-                        </StickyHeader>
-                    )}
+                    <StickyHeader>
+                        {hasDatabaseAdminAccess && <DocumentSchemaSelectActions />}
 
+                        <div className={hasSchemaValidation ? "" : "item-disabled pe-none"}>
+                            {hasAnyValidator && (
+                                <DocumentSchemaFilter
+                                    selectedCollections={selectedCollections}
+                                    setSelectedCollections={handleOnSelectCollection}
+                                    collectionOptions={options}
+                                    selectedStatuses={selectedStatuses}
+                                    setSelectedStatuses={setSelectedStatuses}
+                                    schemasCount={filteredValidators.length}
+                                    isLoading={asyncLoadValidators.loading}
+                                />
+                            )}
+                        </div>
+                    </StickyHeader>
                     <div className="mt-4">
                         <HrHeader
                             count={filteredValidators.length}
@@ -236,7 +128,7 @@ export default function DocumentSchema() {
                                         }}
                                     >
                                         <Button
-                                            size="sm"
+                                            size="xs"
                                             variant="info"
                                             className="rounded-pill"
                                             onClick={handleAddNew}
