+++ src/Raven.Studio/wwwroot/Content/scss/bs5variables.scss
@@ -813,7 +813,7 @@ $zindexes: (
     "text-index": 980,
     "navbar": 1000,
     "sticky-header": 1000,
-    "unpinned-split-view": 1001,
+    "unpinned-split-view": 1101,
     "dropdown": 1002,
     "sticky": 1020,
     "scoped-modal": 1030,
